fix: refactor tweet reply component
