+++ package.json
@@ -12,6 +12,7 @@
     "@typescript-eslint/eslint-plugin": "^5.42.1",
     "@typescript-eslint/parser": "^5.42.1",
     "classnames": "^2.3.2",
+    "dayjs": "^1.11.7",
     "moment": "^2.29.4",
     "next": "12.3.1",
     "react": "18.2.0",
