--- src/components/tweetInfo/style.module.scss
@@ -1,79 +1,0 @@
-.container {
-	height: 100vh;
-	display: flex;
-	flex-direction: column;
-	width: 30%;
-	.tweetCard {
-		display: flex;
-		flex-direction: column;
-		text-align: left;
-		padding: 15px;
-		.tweetCardInfo {
-			display: flex;
-			.tweetCardInfoImg {
-				img {
-					width: 50px;
-					height: 50px;
-					border-radius: 50%;
-					object-fit: cover;
-				}
-			}
-			.tweetCardInfoName {
-				margin-left: 15px;
-				padding: 0;
-				.name {
-					font-size: 15px;
-					font-weight: 700;
-					color: #1c1c1c;
-					letter-spacing: 0.5px;
-				}
-				.account {
-					font-size: 15px;
-					font-weight: 500;
-					color: #657786;
-					letter-spacing: 0.5px;
-				}
-			}
-		}
-		.tweetCardContent {
-			padding: 5px 0;
-			.tweetCardContentText {
-				font-weight: 500;
-				font-size: 23px;
-				color: #1c1c1c;
-				line-height: 34px;
-				margin-top: 0;
-			}
-			.tweetCardContentTime {
-				font-size: 15px;
-				font-weight: 500;
-				color: #657786;
-				letter-spacing: 0.5px;
-				letter-spacing: 0.5px;
-				margin: 10px 0;
-			}
-		}
-		.tweetCardReply {
-			border-top: 1px solid #e6ecf0;
-			padding: 20px 0;
-			span {
-				margin-right: 20px;
-				font-weight: 500;
-				b {
-					color: #1c1c1c;
-					margin: 0 3px;
-				}
-			}
-			svg {
-				display: inline-block;
-				width: 25px;
-				height: 25px;
-				margin: 0 100px 0 3px;
-				cursor: pointer;
-			}
-			&:last-child {
-				padding-bottom: 0;
-			}
-		}
-	}
-}
